html/semantics/embedded-content-0/the-video-element/paused_true_during_pause.html

<!doctype html>
<html>
 <head>
  <title>video events - paused property</title>
  <script src="/resources/testharness.js"></script>
  <script /resources/testharnessreport.js></script>
  <script src="../../../../common/media.js"></script>
 </head>
 <body>
  <p><a href="http://dev.w3.org/html5/spec/Overview.html#mediaevents">spec reference</a></p>
  <video id="v" controls>
  </video>
  <div id="log"></div>
  <script>
var t = async_test("video.paused should be true during pause event", {timeout:30000});
var v = document.getElementById("v");
v.addEventListener("pause", function() {
  t.step(function() {
   assert_true(v.paused);
  });
  t.done();
}, false);
v.src = getVideoURI("/media/movie_300") + "?" + new Date() + Math.random();
v.play();
v.pause();
  </script>
 </body>
</html>
